+++ ChangeLog.md
@@ -4,6 +4,8 @@
 ### [Added]
 - Add `WebMercator` to convert coordinate.
 - Add class `myMap` to get a map from OSM.
+- Add example `testWebMercator` to test `WebMercator`.
+- Add example `testMap` to test the class `myMap`.
 
 ### [Changed]